LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/get-persistendeviceid-from-pointer-mouse-event.tentative-expected.txt


Harness Error (FAIL), message = Error: assert_equals: pointermove deviceId is 1 expected (number) 1 but got (undefined) undefined

PASS PointerEvent.persistentDeviceId
